Added round for bets...

+++ Assets/0_RoulleteProject/4_Scripts/Net/NetData.cs
@@ -18,8 +18,8 @@ public class NetData {
     public int TRANS;
     public float CASH;
 
-    public float SET5;
-    public float SET6;
-    public float SET7;
-    public float SET8;
+    public int SET5;
+    public int SET6;
+    public int SET7;
+    public int SET8;
 }
